requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.114.0
h11==0.14.0
idna==3.8
pillow==10.4.0
pydantic==2.9.1
pydantic_core==2.23.3
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
